lista de productos con disponibilidad

+++ app/Console/Commands/initialize.php
@@ -43,7 +43,7 @@ class initialize extends Command
 
     }
 
-    protected $signature = 'initialize';
+    protected $signature = 'initialize {force?}';
 
     /**
      * The console command description.
@@ -69,32 +69,42 @@ class initialize extends Command
      */
     public function handle()
     {
-        $values = [
-            'PAACK_KEY' => '',
-            'ELINFORMAR_CLIENT_ID' => '',
-            'ELINFORMAR_CLIENT_SECRET' => '',
-            'TWILIO_SID' => '',
-            'TWILIO_TOKEN' => '',
-            'TWILIO_FROM' => '',
-            'TRUUST_PK' => '',
-            'TRUUST_SK' => '',
-            'STRIPE_KEY' => '',
-            'STRIPE_SECRET' => '',
-            'STRIPE_WEBHOOK_SECRET' => '',
-            'STRIPE_WEBHOOK_TOLERANCE' => '',
-            'STRIPE_PLATFORM_ID' => ''
-        ];
-        $this->setEnvironmentValue($values);
+        
+        if ($this->confirm('Quieres continuar?')) {
+            $force = $this->argument('force');
+            
+            $values = [
+                'PAACK_KEY' => '',
+                'ELINFORMAR_CLIENT_ID' => '',
+                'ELINFORMAR_CLIENT_SECRET' => '',
+                'TWILIO_SID' => '',
+                'TWILIO_TOKEN' => '',
+                'TWILIO_FROM' => '',
+                'TRUUST_PK' => '',
+                'TRUUST_SK' => '',
+                'STRIPE_KEY' => '',
+                'STRIPE_SECRET' => '',
+                'STRIPE_WEBHOOK_SECRET' => '',
+                'STRIPE_WEBHOOK_TOLERANCE' => '',
+                'STRIPE_PLATFORM_ID' => ''
+            ];
+            $this->setEnvironmentValue($values);
+            
+            if($force=='force'){
+                Artisan::call('key:generate');
+                Artisan::call('jwt:secret');  
+            }
 
-        Artisan::call('key:generate');
-        Artisan::call('jwt:secret');
-        Artisan::call('migrate:refresh');
-        Artisan::call('db:seed --class=UsersTableSeeder');
-        Artisan::call('db:seed --class=ShopTableSeeder');
-        Artisan::call('db:seed --class=ProductsTableSeeder');
-        Artisan::call('db:seed --class=PurchaseTableSeeder');
-        Artisan::call('db:seed --class=OrderSeeder');
-        Artisan::call('db:seed --class=PurchaseDetailTableSeeder');
-        Artisan::call('db:seed --class=ZoneAvailableTableSeeder');
+            Artisan::call('migrate:refresh');
+            Artisan::call('db:seed --class=UsersTableSeeder');
+            Artisan::call('db:seed --class=ShopTableSeeder');
+            Artisan::call('db:seed --class=ProductsTableSeeder');
+            Artisan::call('db:seed --class=PurchaseTableSeeder');
+            Artisan::call('db:seed --class=OrderSeeder');
+            Artisan::call('db:seed --class=PurchaseDetailTableSeeder');
+            Artisan::call('db:seed --class=ZoneAvailableTableSeeder');
+            Artisan::call('db:seed --class=CategoriesTableSeeder');
+        }
+
     }
 }
